Switch to Decimal from float
Old code crashed as post_fee is a Decimal which could not be multiplied
by price and amount.

+++ cryptotrader/quadrigacx/quadriga_trader.py
@@ -254,11 +254,11 @@ class QuadrigaTrader(Trader):
             if status_code == "2":
                 # Type 0 == Buy, Type 1 == Sell
                 if json_result["type"] == "0":
-                    self.assets = (float(json_result["price"]) * float(json_result["amount"])) * self.post_fee
+                    self.assets = (Decimal(json_result["price"]) * Decimal(json_result["amount"])) * self.post_fee
                     self._active_buy_order = False
                     self._waiting_for_order_to_fill = None
                 else:
-                    self.balance = (float(json_result["price"]) * float(json_result["amount"])) * self.post_fee
+                    self.balance = (Decimal(json_result["price"]) * Decimal(json_result["amount"])) * self.post_fee
                     self._active_sell_order = False
                     self._waiting_for_order_to_fill = None
             elif status_code == "1":
